Fix a critical use-after-free bug
This would cause a crash with shaders active if the BufferBuilder
resized while partway through writing a primitive, as the captured
vertex pointers would then be dangling

It turns out this was also found upstream sometime between Iris 1.7
and 1.8, although the debugging here was done independently. The fix
is borrowed from there.

https://github.com/IrisShaders/Iris/commit/cb049f5ace4bd789a58f563c9d78089dcf4e5ab1